pyproject.toml
[project]
name = "topicos-ia-2024-2do-parcial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.2",
    "llama-index>=0.11.18",
    "llama-index-embeddings-huggingface>=0.3.1",
    "openai>=1.51.2",
    "pydantic-settings>=2.5.2",
    "piccolo[sqlite]>=1.20.0",
]

[tool.uv]
dev-dependencies = [
    "gradio>=5.1.0",
    "ruff>=0.6.9",
]
